+++ src/sagas/channels/vrpc.js
@@ -7,6 +7,7 @@ import {
   SET_WATCHED_VRPC_ADDRESSES,
 } from "../../utils/constants/storeType";
 import VrpcProvider from '../../utils/vrpc/vrpcInterface';
+import Store from '../../store/index';
 
 export default function * vrpcSaga() {
   yield all([
@@ -22,14 +23,35 @@ function * handleVrpcChannelInit(action) {
 }
 
 function* handleVrpcChannelClose(action) {
-  VrpcProvider.deleteEndpoint(
-    action.payload.systemId,
-    action.payload.endpointAddress,
-  );
+  try {
+    // Check if the endpoint exists before trying to delete it
+    const endpoints = Store.getState().channelStore_vrpc.vrpcEndpoints;
+    const endpointId = VrpcProvider.getEndpointId(action.payload.systemId, action.payload.endpointAddress);
+    
+    if (endpoints[endpointId]) {
+      VrpcProvider.deleteEndpoint(
+        action.payload.systemId,
+        action.payload.endpointAddress,
+      );
+    } else {
+      console.log(`VRPC endpoint ${action.payload.endpointAddress} already deleted for ${action.payload.systemId}`);
+    }
+  } catch (error) {
+    // Endpoint might not be initialized, which is fine during cleanup
+    console.log('VRPC channel close:', error.message);
+  }
 }
 
 function * handleSignOut() {
-  VrpcProvider.deleteAllEndpoints();
+  // Only delete endpoints if they haven't been deleted already
+  try {
+    const endpoints = Store.getState().channelStore_vrpc.vrpcEndpoints;
+    if (Object.keys(endpoints).length > 0) {
+      VrpcProvider.deleteAllEndpoints();
+    }
+  } catch (error) {
+    console.log('VRPC sign out cleanup:', error.message);
+  }
   
   setImmediate(() => {
     VrpcProvider.addDefaultEndpoints();
